chore: finish api refactor
